did a lot of work with an AuthGuard
I made and Authorization Guard and got it working with a fake backend. It is functioning correctly. Need to connect it will backend. Working on registration and already finished logout!!! nav bar is a little crazy and im going to get it all looking like an actual website next! super cool stuff!